pyproject.toml
[project]
name = "ai-soc-analyst"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.3",
    "pydantic-ai[all]>=0.8.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
